+++ nibRealUI/Core/AddonData/AddonData.lua
@@ -5,6 +5,7 @@ local next = _G.next
 
 -- RealUI --
 local RealUI = private.RealUI
+local L = RealUI.L
 
 RealUI.AddOns = {}
 function RealUI:LoadAddonData()
@@ -13,12 +14,36 @@ function RealUI:LoadAddonData()
     end
 end
 
+RealUI.Profiles = {}
+function RealUI:LoadAddonProfiles()
+    for name, func in next, self.Profiles do
+        func()
+    end
+end
+
 function RealUI:LoadSpecificAddOnData(addon, skipReload)
     --print("RealUI:LoadSpecificAddOnData", addon, skipReload, self.AddOns[addon])
     if self.AddOns[addon] then
         self.AddOns[addon]()
-        --setProfile
-        if skipReload then return end
-        self:ReloadUIDialog()
     end
+    if self.Profiles[addon] then
+        self.Profiles[addon]()
+    end
+
+    if skipReload then return end
+    self:ReloadUIDialog()
 end
+
+_G.StaticPopupDialogs["RealUI_ResetAddonProfile"] = {
+    text = L["Patch_UpdateAddonSettings"],
+    button1 = _G.YES,
+    button2 = _G.NO,
+    OnAccept = function(self)
+        RealUI:LoadSpecificAddOnData(self.text.text_arg1)
+    end,
+    OnCancel = function() end,
+    timeout = 0,
+    whileDead = true,
+    hideOnEscape = false,
+    notClosableByLogout = false,
+}